+++ lib/Query/QueryBuilder.php
@@ -14,7 +14,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  *
- * @version    0.2.0
+ * @version    0.3.0
  * @copyright  2017-2020 Kristuff
  */
 
@@ -25,6 +25,7 @@ use Kristuff\Patabase\Query\QueryBase;
 use Kristuff\Patabase\Exception;
 use Kristuff\Patabase\SqlException;
 use Kristuff\Patabase\Driver;
+use Kristuff\Patabase\Outpout;
 
 /**
  * Class QueryBuilder
@@ -294,23 +295,24 @@ abstract class QueryBuilder extends QueryBase
      * 
      * @return mixed                     
      */
-    protected static function fetchOutput(QueryBuilder $query, $executed, $ouputFormat)
+    protected static function fetchOutput(QueryBuilder $query, $executed, $outputFormat)
     {
-        switch ($ouputFormat){
-            case 'asso':    
+        switch (strtoupper($outputFormat)){
+
+            case Outpout::ASSOC:    
                 return $executed ? $query->pdoStatement->fetchAll(\PDO::FETCH_ASSOC) :  array();
 
-            case 'obj':     
-                return $executed ? $query->pdoStatement->fetchAll(\PDO::FETCH_OBJ) :    NULL;
+            case Outpout::OBJ:    
+                return $executed ? $query->pdoStatement->fetchAll(\PDO::FETCH_OBJ) :    array();
 
-            case 'cols':     
+            case Outpout::COLUMN:    
                 return $executed ? $query->pdoStatement->fetchAll(\PDO::FETCH_COLUMN) :  array();
 
-            case 'json':    
+            case Outpout::JSON:
                 $results = $executed ? $query->pdoStatement->fetchAll(\PDO::FETCH_ASSOC) :  array();
                 return json_encode($results, JSON_NUMERIC_CHECK);   
 
-            case 'jsonpp':    
+            case Outpout::JSON_PRETTY_PRINT:    
                 $results = $executed ? $query->pdoStatement->fetchAll(\PDO::FETCH_ASSOC) :  array();
                 return json_encode($results, JSON_PRETTY_PRINT | JSON_NUMERIC_CHECK);