+++ lib/Query/SelectBase.php
@@ -14,7 +14,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  *
- * @version    0.2.0
+ * @version    0.3.0
  * @copyright  2017-2020 Kristuff
  */
 
@@ -180,7 +180,7 @@ abstract class SelectBase extends QueryBuilder
     }
 
     /**
-     * Define the ouputs columns for the select
+     * Define the outputs columns for the select
      *
      * @access public
      * @return $this